sysext.just: Setup SELinux modules manually
Load the SELinux modules as needed in the build container to make sure
custom labels get set for files in the sysext.

The SELinux policy modules are not setup in the final system yet, when
the sysext is merged.

Taken from: https://src.fedoraproject.org/rpms/selinux-policy/blob/rawhide/f/rpm.macros