sysext.just: Setup SELinux modules manually
Load the SELinux modules as needed in the build container to make sure
custom labels get set for files in the sysext.

Note that the SELinux policy modules are not yet set up in the final
system, when the sysext is merged.

Taken from: https://src.fedoraproject.org/rpms/selinux-policy/blob/rawhide/f/rpm.macros