sysext.just: Auto merge /usr/sbin with /usr/bin for F42+
Fedora packages should have been fixed for F42 but external packages may
or not have been fixed yet. Do this automatically as much as possible as
there is little value in failling on this in CI on package changes.

Update the comment for the validate step to better reflect what it is
doing.